requirements.txt
asgiref==3.7.2
Django==4.2.2
django-crispy-forms==1.14.0
django-environ==0.10.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
phonenumbers==8.13.18
Pillow==9.5.0
python-magic==0.4.27
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
whitenoise==6.5.0
